refactor(binance): move BinanceMarketStreamAdapter and BinanceUserStreamAdapter to adapter-binance
Classes implementing core ports must live in the adapter module, not in the
Spring composition layer. Spring configurations remain unchanged — they only
wire the beans.

Also adds .ai/ protocol directory with architecture, coding standards,
validation, review, change-safety, and docs guides.
