+++ .ai/decisions/README.md
@@ -1,0 +1,24 @@
+# Decisoes Arquiteturais
+
+Este diretorio existe para registrar decisoes arquiteturais curtas e duraveis.
+
+## Quando criar uma decisao
+
+- quando uma regra importante nao for obvia
+- quando houver tradeoff arquitetural relevante
+- quando uma excecao controlada precisar ser documentada
+
+## Convencao sugerida
+
+Nomear arquivos como:
+
+- `001-nome-curto-da-decisao.md`
+- `002-outra-decisao.md`
+
+## Estrutura sugerida
+
+- contexto
+- decisao
+- impacto
+- alternativas descartadas
+
